Add files via upload
Blood Bank Application:
The Mobile App Blood Bank system aims to save lives by providing easy access to information about blood donors and their medical details. Users can quickly check blood availability and register as donors after logging in. The system relies on internet access, which is its main limitation, but it significantly speeds up the process of finding the right donor for those in need.

Features:
 User login:
This allows only the registered users to login in order to use this location tracking application.
 Donor details:
This module helps the donor to insert all the necessary details that is personal information and
medical information plus the blood group which together helps to select a donor.
 Procurer details:
The mobile app user can view all the donor details and select the required donor on the basis of
the donor’s information.
 Front End:
The React Stack: ReactNative, Redux, React Navigation, Native Base
 Backend: Firebase